requirements/dev.txt
# Local development dependencies go here
-r base.txt
django-debug-toolbar==1.3.0
django-extensions==1.5.5
Werkzeug==0.10.4

